Merge branch 'log-project-deletion' into 'master'
Add a log message when a project is scheduled for destruction for debugging

We have a lot of projects that are in `pending_delete` state. It's not clear whether they were ever scheduled for destruction, or whether Sidekiq just dropped the job due to `MemoryKiller` or some other reason.

Also this will provide a record of which user destroys a project.

#20365

See merge request !5540

Signed-off-by: Rémy Coutable <remy@rymai.me>