Minor update on CI docs examples
Triggered by gitlab-org/gitlab-ci-yml#1 I've investigated how this is
organized so far and did a minor clean up. In my opinion we should just
keep a list of up to date blog posts in the docs and examples should be
included in gitlab-org/gitlab-ci-yml. This reduces the number of
projects and docs pages one has to look for information.

[ci skip]