Allow people to subscribe to mentions in updated MRs and Issues
This slightly changes the semantics of the 'New Issue' and 'New MR' events to
include new mentions in edited Mentionables.

An alternative would be to introduce 'Issue updated' and 'MR updated' events,
but that would lead to questions about why those events were only available
to new mentions, and not existing mentions as well, so hold off for now.