Merge branch 'fix-empty-dropdown' into 'master'
Fix empty dropdown

Deselecting the current selected option on a dropdown they were removing the input related to them.
This led to a unexpected cases. The forms were submitted without all required inputs and the label was still indicating an option was selected. This MR fixes that.

See merge request !5927