Merge branch '21378-error-500-after-8-11-update-on-system-infos-page' into 'master'
Handle unavailable system info

## What does this MR do?

Handle the case where we can't get system info without blowing up. As this is the first tab in the monitoring section, it's difficult to get to the other tabs if this page throws a 500. Also be more specific about the info we want, so we don't fail on something we don't care about (like `/proc/net/dev`).

## Why was this MR needed?

grsecurity can prevent users from reading `/proc`, which is what Vmstat uses to find CPU and memory info.

## What are the relevant issue numbers?

Closes #21378.

See merge request !5989