Expire commit view partial after a day
We rarely use Russian-doll caching in views, and when we do, it's
typically with a naturally-invalidating key. In the case of the commit
partial, though, the author lookup isn't part of the cache key - because
we're not going to use the state of the users table - and so a new email
address can take up to two weeks to show against the commits list.

Limiting this to a day still caches the partial for a healthy amount of
time, without as bad a worst case scenario.