Append issue template to existing description
Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/21733

Add two newlines before the template if the existing description isn't
empty. This makes it easier to see where the template begins.

Don't append the template when "Reset template" is selected, of course.

Don't append template if it equals the existing description. This makes
it so that selecting a template twice doesn't duplicate it.