Fix migration for removing MR diff indexes
For whatever reason on some PostgreSQL installations there would be a
separate UNIQUE constraint on the merge_request_id column. Rails'
index_exists?() returns false for this constraint (even when using the
full name), yet the indexes() method returns it.

This commit changes the migration so that the constraint is dropped as
well. MySQL installations don't appear to be affected.

Fixes gitlab-org/gitlab-ce#22136