Improve validity of spec/features/todos/todos_filtering_spec.rb
Previously, we were checking that a CSS selector string didn't have some
page-related content, which, of course it didn't.

Now we check against the actual content of the selector, we use a more
semantic selector, and we add an additional expectation for the text
that _should_ be there, as an additional sanity check.