Merge branch '22020-fix-form-submit-missing-url' into 'master'
Add missing url param to ajax call

The `url` param was missing from the ajax call.

This bug happened because the `url` param was not set. jQuery defaults to `location.href`. BUT if the user came to the protected branches page from another page a previously set `ajaxSettings.url` was used as default url (which defaults to `location.href`) so when the user submitted the form the request was made to a  completely wrong `url`. Refreshing the page was working because `ajaxSettings.url` was set to the current page'`location.href` and that was the correct URL.

This MR ensures the url will always be set with the right one.

Fixes https://gitlab.com/gitlab-org/gitlab-ce/issues/22020

See merge request !760