Ensure absolute URLs for single lines from Banzai for HipChat
"pipeline: :single_line" leaves the protocol/host part out of the URLs
and caches them that way. To avoid giving those out to HipChat, markdown
is always rendered with "pipeline: :email" first.

There must be a better way to do this, but I can't see how to avoid the
link caching.